ParticleGUI/pyproject.toml
[project]
name = "particlegui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ncca-ngl>=0.3.4",
]
